% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtChart.R
\name{RtChart}
\alias{RtChart}
\title{R chart function
This function allows you to create a chart displaying the reproduction number over time}
\usage{
RtChart(ctr)
}
\arguments{
\item{country}{\itemize{
\item specify the country to draw the chart for
RtChart()
}}
}
\description{
R chart function
This function allows you to create a chart displaying the reproduction number over time
}
